simplify pcr reader logic
